refactor: extract class wrapper from ruby-generator index
Move _wrapWithClass, _isValidClassName, _generateSetXxx, and
_generateStageSetXxx from index.js to class-wrapper.js.

Part of #302
